requirements.txt
absl-py==0.4.1
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
bleach==1.5.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.5
colorama==0.3.9
cycler==0.10.0
Cython==0.28.5
dask==0.19.0
decorator==4.3.0
deepdish==0.3.6
enum34==1.1.6
flatbuffers==1.11
funcsigs==1.0.2
future==0.16.0
gast==0.2.0
getch==1.0
grpcio==1.15.0
h5py==2.8.0
html5lib==0.9999999
idna==2.7
imageio==2.4.1
imutils==0.5.1
iso8601==0.1.12
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.3
more-itertools==4.3.0
moviepy==0.2.3.5
networkx==2.1
numexpr==2.6.8
numpy==1.14.5
numpy-stl==2.10.1
opencv-python==3.4.3.18
pathlib2==2.3.2
Pillow==5.2.0
pluggy==0.7.1
protobuf==3.6.1
psutil==5.4.7
py==1.6.0
pycparser==2.18
pyparsing==2.2.0
pyquaternion==0.9.2
pyserial==3.4
pytest==3.8.0
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.5
PyWavelets==1.0.0
PyYAML==3.13
redis==2.10.6
requests==2.19.1
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
stl==0.0.3
tensorboard==1.10.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
toolz==0.9.0
tqdm==4.26.0
urllib3==1.23
Werkzeug==0.14.1
omegaconf==1.4.1